gptp: Update shared memory with reverse sync config file parameter

+++ examples/libgptp_test/libgptp_test.cpp
@@ -172,7 +172,7 @@ void loop_test(int time_us)
     int64_t delta_qtimer_time;
     int64_t delta_ptp_time;
     int64_t delta_qtimer_ptp;
-    int16_t time_error;
+    int64_t time_error;
     int rcvid = 0;
     prev_qtimer_time = getQtimerTime();
     gptpGetCurPtpTime(&prev_ptp_time);
